requirements.txt
chromadb
langchain
langchain-community
langchain-openai
python-dotenv
requests
streamlit